# run-tests: add skip cache to improve performance