@nikic 2021-02-24
There's a failure of ext/mysqli/tests/mysqli_pam_sha256_public_key_option_invalid.phpt on some jobs that I think is related to this change (https://dev.azure.com/phpazuredevops/PHP/_build/results?buildId=14943&view=ms.vss-test-web.build-test-results-tab&runId=340192&paneView=debug&resultId=107434).