@MaxSem 2021-02-24
@nikic, since tests are running faster now, the race condition with various mysqli tests recreating the same table is more likely to occur. Also, this extension's tests frequently mix skip checks with database setup for tests, making it possible to run a test without setting up tables for it, but that's not a cause for that particular failure because the test has unique skipif code.

I'll try to address these issues soon.