@nikic 2021-02-24
@MaxSem mysqli tests don't run in parallel -- there shouldn't be race conditions there.