@MaxSem 2021-02-24
Hmm, how 'bout we put a temporary plug to avoid caching mysqli tests to see if it fixes nightlies? Meanwhile, these tests need fixing anyway to even work with recent mysql...